Added Clustering for Lab1, Added Files Lab2
For Lab1, I added kmeans clustering to find the dominant colors of a
frame of my video capture. I essentially do kmeans and then output the
most dominant colors onto a palette.
I also started on Lab2 adding the subscriber and publisher code for
MQTT. The code allows for a subscriber to read messages from a publisher
that publishes their message onto a "topic." Only members of a "topic"
can read and write to it.