Added Gesture Classification Code
Added the code my group used to identify different gestures and to check
for error. We fed training data to our decesion tree classifier to train
it to identify what gestures were performed. Then we fed testing data to
see how well our classification performed. It did not perform with 100%
accuracy so more changes will need to be made.